requirements.txt
# Use CPU-only PyTorch wheels (smaller install size)
--extra-index-url https://download.pytorch.org/whl/cpu

numpy
sentence-transformers
torch==2.4.0
pysolr
fireworks-ai
pydantic>=2
pyyaml
fastapi
uvicorn
python-dotenv
httpx